Ch25: add file system related code
In the original code, MINT64OS boots from a floppy disk and the
filesystem code is set to format and mount the primary master IDE disk,
which means users cannot set which disk to format and mount.

But I modified the code because my MINT64OS boots from the primary
master hard disk. The modified filesystem code is set up to format and
mount the primary secondary IDE disk.